+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added the ability to create a custom form based on the model schema via `linz.api.model.generateForm()` and `linz.api.model.generateFormString()`.
+- Added the ability to define custom fields in the form DSL. When specifying a custom field you should add a `type` property if you want an input other than `text`.
 - Add the ability to edit `public` and `views` files without restarting the node process.
 - Replaced `linz.publicMiddleware.namespace` with `linz.api.middleware.setLinzNamespace()`.
 - Added persistent storage to the local development environment.
@@ -63,8 +64,6 @@
 
 ### Improvements
 
-- Added `model.generateForm(model, options)` and `model.generateFormString(model, options)` API functions to allow generating a form using a Linz form DSL.
-- Added the ability to define custom fields in the form DSL. When specifying a custom field you should add a `type` property if you want an input other than `text`.
 - Added support for a `content` property when supplying `scripts` and `styles` defaults. Supplying `content` for a script will ensure the content is placed within two script tags (i.e. `<script>{{{content}}}</script>`). Supplying `content` for a style will ensure the content is placed within two style tags (as opposed to using a link tag) (i.e. `<style>{{{content}}}</style>`).
 - Added the `linz.formtools.cellRenderers.email` renderer to render an email in an `a` with the `mailto:` protocol.
 - Added the `linz.formtools.cellRenderers.tel` renderer to render a telephone number in an `a` with the `tel:` protocol.
